+++ src/serpent/settings.py
@@ -1,5 +1,7 @@
-# TODO Read config from a `.serpentrc` file
+"""Serpent settings."""
 from __future__ import annotations
+
+# TODO Read settings from a `.serpentrc` file
 
 BASE_ORDER = 'GACT'
 COUNT_LIMIT = 20
